+++ src/sentry/models/organizationmember.py
@@ -23,7 +23,7 @@ from sentry.db.models import (
     BoundedPositiveIntegerField,
     FlexibleForeignKey,
     Model,
-    control_silo_only_model,
+    region_silo_only_model,
     sane_repr,
 )
 from sentry.db.models.manager import BaseManager
@@ -99,7 +99,7 @@ class OrganizationMemberManager(BaseManager):
         return user_teams
 
 
-@control_silo_only_model
+@region_silo_only_model
 class OrganizationMember(Model):
     """
     Identifies relationships between organizations and users.
