+++ tests/sentry/api/endpoints/test_accept_organization_invite.py
@@ -14,10 +14,11 @@ from sentry.models import (
     OrganizationMember,
 )
 from sentry.testutils import TestCase
+from sentry.testutils.factories import Factories
 from sentry.testutils.silo import control_silo_test, exempt_from_silo_limits
 
 
-@control_silo_test(stable=True)
+@control_silo_test()
 class AcceptInviteTest(TestCase):
     def setUp(self):
         super().setUp()
@@ -45,7 +46,8 @@ class AcceptInviteTest(TestCase):
         return reverse(url, args=[self.organization.slug] + args)
 
     def _require_2fa_for_organization(self):
-        self.organization.update(flags=F("flags").bitor(Organization.flags.require_2fa))
+        with exempt_from_silo_limits():
+            self.organization.update(flags=F("flags").bitor(Organization.flags.require_2fa))
         assert self.organization.flags.require_2fa.is_set
 
     def _assert_pending_invite_details_in_session(self, om):
@@ -69,17 +71,16 @@ class AcceptInviteTest(TestCase):
             assert resp.status_code == 400
 
     def test_invalid_token(self):
-        with exempt_from_silo_limits():
-            om = OrganizationMember.objects.create(
-                email="newuser@example.com", token="abc", organization=self.organization
-            )
+        om = Factories.create_member(
+            email="newuser@example.com", token="abc", organization=self.organization
+        )
         for path in self._get_paths([om.id, 2]):
             resp = self.client.get(path)
             assert resp.status_code == 400
 
     def test_invite_not_pending(self):
         user = self.create_user(email="test@gmail.com")
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", token="abc", organization=self.organization, user=user
         )
         for path in self._get_paths([om.id, om.token]):
@@ -87,7 +88,7 @@ class AcceptInviteTest(TestCase):
             assert resp.status_code == 400
 
     def test_invite_unapproved(self):
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com",
             token="abc",
             organization=self.organization,
@@ -98,7 +99,7 @@ class AcceptInviteTest(TestCase):
             assert resp.status_code == 400
 
     def test_needs_authentication(self):
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", token="abc", organization=self.organization
         )
         for path in self._get_paths([om.id, om.token]):
@@ -109,7 +110,7 @@ class AcceptInviteTest(TestCase):
     def test_not_needs_authentication(self):
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", token="abc", organization=self.organization
         )
         for path in self._get_paths([om.id, om.token]):
@@ -123,7 +124,7 @@ class AcceptInviteTest(TestCase):
 
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", token="abc", organization=self.organization
         )
 
@@ -140,7 +141,7 @@ class AcceptInviteTest(TestCase):
 
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", token="abc", organization=self.organization
         )
         for path in self._get_paths([om.id, om.token]):
@@ -154,7 +155,7 @@ class AcceptInviteTest(TestCase):
         AuthProvider.objects.create(organization=self.organization, provider="google")
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", token="abc", organization=self.organization
         )
         for path in self._get_paths([om.id, om.token]):
@@ -171,7 +172,7 @@ class AcceptInviteTest(TestCase):
             user = self.create_user(f"boo{i}@example.com")
             self.login_as(user)
 
-            om = OrganizationMember.objects.create(
+            om = Factories.create_member(
                 email=user.email,
                 role="member",
                 token="abc",
@@ -197,7 +198,7 @@ class AcceptInviteTest(TestCase):
     def test_cannot_accept_expired(self):
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", token="abc", organization=self.organization
         )
         OrganizationMember.objects.filter(id=om.id).update(
@@ -215,7 +216,7 @@ class AcceptInviteTest(TestCase):
     def test_cannot_accept_unapproved_invite(self):
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com",
             role="member",
             token="abc",
@@ -238,7 +239,7 @@ class AcceptInviteTest(TestCase):
             user = self.create_user(f"boo{i}@example.com")
             self.login_as(user)
 
-            om = OrganizationMember.objects.create(
+            om = Factories.create_member(
                 email=user.email,
                 role="member",
                 token="abc",
@@ -252,7 +253,7 @@ class AcceptInviteTest(TestCase):
             assert om.email is None
             assert om.user == user
 
-            om2 = OrganizationMember.objects.create(
+            om2 = Factories.create_member(
                 email="newuser3@example.com",
                 role="member",
                 token="abcd",
@@ -273,7 +274,7 @@ class AcceptInviteTest(TestCase):
 
             self.login_as(user)
 
-            om = OrganizationMember.objects.create(
+            om = Factories.create_member(
                 email="newuser" + str(i) + "@example.com",
                 role="member",
                 token="abc",
@@ -305,7 +306,7 @@ class AcceptInviteTest(TestCase):
 
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com", role="member", token="abc", organization=self.organization
         )
         for path in self._get_paths([om.id, om.token]):
@@ -322,7 +323,7 @@ class AcceptInviteTest(TestCase):
 
             self.login_as(user)
 
-            om = OrganizationMember.objects.create(
+            om = Factories.create_member(
                 email="newuser" + str(i) + "@example.com",
                 role="member",
                 token="abc",
@@ -342,7 +343,7 @@ class AcceptInviteTest(TestCase):
     def test_mismatched_org_slug(self):
         self.login_as(self.user)
 
-        om = OrganizationMember.objects.create(
+        om = Factories.create_member(
             email="newuser@example.com",
             role="member",
             token="abc",
