+++ tests/sentry/api/endpoints/test_assistant.py
@@ -4,8 +4,10 @@ from exam import fixture
 from sentry.assistant import manager
 from sentry.models import AssistantActivity
 from sentry.testutils import APITestCase
+from sentry.testutils.silo import control_silo_test
 
 
+@control_silo_test(stable=True)
 class AssistantActivityTest(APITestCase):
     endpoint = "sentry-api-0-assistant"
 
@@ -45,6 +47,7 @@ class AssistantActivityTest(APITestCase):
         assert {"guide": guide, "seen": True} in resp.data
 
 
+@control_silo_test(stable=True)
 class AssistantActivityUpdateTest(APITestCase):
     endpoint = "sentry-api-0-assistant"
     method = "put"
