+++ tests/sentry/api/endpoints/test_auth_config.py
@@ -4,10 +4,10 @@ from django.test.utils import override_settings
 
 from sentry import newsletter
 from sentry.testutils import APITestCase
-from sentry.testutils.silo import control_silo_test
+from sentry.testutils.silo import control_silo_test, exempt_from_silo_limits
 
 
-@control_silo_test
+@control_silo_test(stable=True)
 class AuthConfigEndpointTest(APITestCase):
     path = "/api/0/auth/config/"
 
@@ -29,6 +29,7 @@ class AuthConfigEndpointTest(APITestCase):
         assert response.data["nextUri"] == "/organizations/ricks-org/issues/"
 
     @override_settings(SENTRY_SINGLE_ORGANIZATION=True)
+    @exempt_from_silo_limits()  # Single org IS monolith mode
     def test_single_org(self):
         response = self.client.get(self.path)
